requirements.txt
Django==1.8.5
django-overextends==0.4.0
django-phonenumber-field==0.7.2
gunicorn==19.3.0
phonenumbers==7.1.0
Pillow==3.0.0
sphinx-me==0.3
wheel==0.24.0
